jquery friend add, remove, list

+++ app/assets/javascripts/show.js
@@ -3,24 +3,68 @@ $(document).ready(function(){
     url: '/friendships/1.json',
     dataType: 'JSON',
     method: 'GET',
-    success: function(data){
-      if(data.length){
-        items = "<div><h3>Accept</h3></div> ";
-        $('.friends_list').html(items);
+    success: succesFunction
+  });
+
+  function succesFunction(data) {
+    if(data.friendships.length > 0){
+      var user = $.map(data.users, function(user){
+        return user;
+      });
+      var friends = $.map(data.friendships, function(fr){
+        return fr;
+      });
+      for(var i in user) {
+        for(var j in friends) {
+          var hrefPage = window.location.href;
+          var idPage = hrefPage.replace('http://localhost:3000/friendships/', ' ');
+
+          if(user[i].id === friends[j].user_id && friends[j].friend_id == idPage){
+            console.log("Кто принял заявку");
+
+            var user_name = user[i].first_name + " " + user[i].last_name;
+            var email = "<strong>Email:</strong>" + " " + user[i].email;
+            var avatar = "<img class='thumb media-object' src='" + user[i].avatar['url'] + "'>";
+
+            $('h4.friends-name').html(user_name);
+            $('address.friend-address').html(email);
+            $('.pull-left__friend').html(avatar);
+
+            $('.media--friend_decline').children().on('click', function(){
+              $.ajax({
+                url: "/friendships/" + friends[j].user_id,
+                method: 'DELETE',
+                success: function(){
+                  console.log("Succes");
+                }
+              });
+            });
+
+            $('.media--friend_accept').children().on('click', function(){
+              $.ajax({
+                url: "/friendships/" + friends[j].user_id,
+                method: 'PUT',
+                success: function(){
+                  console.log("Succes");
+                }
+              });
+            });
+          }
+          else if(user[i].id === friends[j].friend_id && friends[j].user_id == idPage){
+            console.log("Кто отправил заявку");
+            var user_name = user[i].first_name + " " + user[i].last_name;
+            var email = "<strong>Email:</strong>" + " " + user[i].email;
+            var avatar = "<img class='thumb media-object' src='" + user[i].avatar['url'] + "'>";
+
+            $('h4.friends-name').html(user_name);
+            $('address.friend-address').html(email);
+            $('.pull-left__friend').html(avatar);
+          }
+        }       
       }
     }
-  });
-});
-
-$(document).ready(function(){
-  var url = document.location.href;
-  $('.decline_friend').on('click', function(){
-    $.ajax({
-      url: url,
-      method: 'DELETE',
-      success: function(){
-        console.log("Succes");
-      }
-    });
-  });
+    else{
+      $('.panel__friend').remove();
+    }
+  }
 });
